+++ src/main/java/com/github/johnnyjayjay/discord/commandapi/CommandListener.java
@@ -32,8 +32,11 @@ class CommandListener extends ListenerAdapter {
             if (raw.startsWith(prefix)) {
                 long timestamp = System.currentTimeMillis();
                 long userId = event.getAuthor().getIdLong();
-                if (cooldowns.containsKey(userId) && (timestamp - cooldowns.get(userId)) < settings.getCooldown())
+                if (cooldowns.containsKey(userId) && (timestamp - cooldowns.get(userId)) < settings.getCooldown()) {
+                    if (settings.isResetCooldown())
+                        cooldowns.put(userId, timestamp);
                     return;
+                }
                 cooldowns.put(userId, timestamp);
                 CommandEvent.Command cmd = new CommandEvent.Command(raw, prefix, settings);
                 if (cmd.getExecutor() != null) {